% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{calcium}
\alias{calc.rho}
\alias{calcium}
\alias{calciumse}
\alias{fastgluc}
\alias{fastglucse}
\title{Data on effect of calcium on fasting glucose (correlated variants)}
\format{An object of class \code{numeric} of length 6.}
\usage{
calcium

calciumse

fastgluc

fastglucse

calc.rho
}
\description{
Two sets of example data are included in the package: one illustrating uncorrelated variants, and the other correlated variants. These are the data on correlated variants.
}
\details{
The variables \code{calcium}, and \code{fastgluc} are the genetic associations with calcium and fasting glucose for 7 genetic variants reported by Burgess et al (2015). The respective standard errors of the associations are given as \code{calciumse} and \code{fastglucse}. The matrix of correlations between the genetic variants is given as \code{calc.rho}.

These data can be used to test out the various functions in the package.
}
\references{
Stephen Burgess, Robert A Scott, Nic J Timpson, George Davey Smith, Simon G Thompson. Using published data in Mendelian randomization: a blueprint for efficient identification of causal risk factors. Eur J Epidemiol 2015; 30(7):543-552. doi: 10.1007/s10654-015-0011-z.
}
\keyword{datasets}

