% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtraFunctions.R
\name{ci_normal}
\alias{ci_normal}
\title{Calculate confidence intervals using the normal distribution}
\usage{
ci_normal(type, mean, se, alpha)
}
\arguments{
\item{type}{"l" for lower, "u" for upper.}

\item{mean}{Causal estimate.}

\item{se}{Standard error of estimate.}

\item{alpha}{Significance level.}
}
\value{
Numeric value of confidence interval limit.
}
\description{
Internal function for calculating confidence intervals using the normal distribution.
}
\details{
None.
}
\examples{
ci_normal(type = "l", mean = 0, se = 1, alpha = 0.05)

}
\keyword{internal}
