% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtraFunctions.R
\name{ci_t}
\alias{ci_t}
\title{Calculate confidence intervals using the t-distribution}
\usage{
ci_t(type, mean, se, df, alpha)
}
\arguments{
\item{type}{"l" for lower, "u" for upper.}

\item{mean}{Causal estimate.}

\item{se}{Standard error of estimate.}

\item{df}{Number of degrees of freedom.}

\item{alpha}{Significance level.}
}
\value{
Numeric value of confidence interval limit.
}
\description{
Internal function for calculating confidence intervals using the t-distribution.
}
\details{
None.
}
\examples{
ci_t(type = "l", mean = 0, se = 1, df = 8, alpha = 0.05)

}
\keyword{internal}
