% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CLR-class}
\alias{CLR-class}
\title{CLR Class}
\description{
An object containing the confidence intervals produced using the conditional likelihood ratio method (CLR) and other identification-robust methods.
}
\section{Slots}{

\describe{
\item{\code{Exposure}}{The name of the exposure variable.}

\item{\code{Outcome}}{The name of the outcome variable.}

\item{\code{Correlation}}{The matrix of correlations between genetic variants.}

\item{\code{ARlower}}{The lower bounds of the causal estimate based on inverting Anderson and Rubin's test.}

\item{\code{ARupper}}{The upper bounds of the causal estimate based on inverting Anderson and Rubin's test.}

\item{\code{Klower}}{The lower bounds of the causal estimate based on inverting Kleibergen's test.}

\item{\code{Kupper}}{The upper bounds of the causal estimate based on inverting Kleibergen's test.}

\item{\code{CLRlower}}{The lower bounds of the causal estimate based on inverting Moreira's conditional likelihood ratio test.}

\item{\code{CLRupper}}{The upper bounds of the causal estimate based on inverting Moreira's conditional likelihood ratio test.}

\item{\code{CIMin}}{The smallest value used in the search to find the confidence interval.}

\item{\code{CIMax}}{The largest value used in the search to find the confidence interval.}

\item{\code{CIStep}}{The step size used in the search to find the confidence interval.}

\item{\code{Alpha}}{The significance level used in constructing the confidence interval (default is 0.05).}
}}

