% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_mvivw-methods.R
\name{condFstat}
\alias{condFstat}
\title{Calculates conditional F-statistic for each risk factor using summarized data}
\usage{
condFstat(.Bx, .Bxse, .nx, .ld, .cor.x)
}
\arguments{
\item{.Bx}{Beta-coefficient for genetic associations with the risk factors.}

\item{.Bxse}{Standard error of genetic associations with the risk factors.}

\item{.nx}{sample sizes used to compute genetic associations with the risk factors.}

\item{.ld}{genetic variant correlation matrix}

\item{.cor.x}{Risk factor correlation matrix}
}
\value{
Conditional F-statistic for each exposure
}
\description{
Internal function for calculating conditional F-statistics for the output \code{condFstat}.
}
\details{
None.
}
\examples{
condFstat(.Bx = cbind(ldlc, hdlc, trig), .Bxse = cbind(ldlcse, hdlcse, trigse),
 .nx = rep(17723,3), .ld=diag(length(ldlc)), .cor.x=diag(3))

}
\keyword{internal}
