% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{coursedata}
\alias{coursedata}
\title{Course data}
\format{An object of class \code{data.frame} with 1000 rows and 7 columns.}
\usage{
coursedata
}
\description{
Data required for Mendelian randomization course.
}
\keyword{internal}
