% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtraFunctions.R
\name{decimals}
\alias{decimals}
\title{Produce nicely rounded numbers}
\usage{
decimals(number, places)
}
\arguments{
\item{number}{A number to be rounded.}

\item{places}{Number of decimal places.}
}
\value{
Returns a nicely formatted number. For example, 0.3 to three decimal places would be rendered as 0.300.
}
\description{
Internal function for putting extra zeros on numbers (if needed).
}
\details{
None.
}
\examples{
decimals(number = 0.3, places = 3)

}
\keyword{internal}
