% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_hetpen-methods.R
\name{het.weight}
\alias{het.weight}
\title{Heterogeneity-penalized weight function}
\usage{
het.weight(prob.valid.inst, bx, by, byse)
}
\arguments{
\item{prob.valid.inst}{Probability (prior) of a genetic variant being a valid instrument.}

\item{bx}{Genetic associations with risk factor.}

\item{by}{Genetic associations with outcome.}

\item{byse}{Standard errors of genetic associations with outcome.}
}
\value{
Heterogeneity-penalized weight.
}
\description{
Internal function for calculating heterogeneity-penalized weights.
}
\details{
None.
}
\examples{
het.weight(0.5, calcium, fastgluc, fastglucse)

}
\keyword{internal}
