% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_mvcML-methods.R
\name{invcov_mvmr}
\alias{invcov_mvmr}
\title{Generate the list of inverse of covariance matrices used in \code{MVMR-cML-DP}}
\usage{
invcov_mvmr(se_bx, se_by, rho_mat)
}
\arguments{
\item{se_bx}{A m*L matrix of standard errors of SNP-exposure association}

\item{se_by}{A vector of standard errors of SNP-outcome association}

\item{rho_mat}{The correlation matrix among the L exposures and the outcome}
}
\value{
A list of inverse of covariance matrices with respect to each genetic variant, retaining the ordering in \code{se_bx}
}
\description{
Generate the list of inverse of covariance matrices used in \code{MVMR-cML-DP}
}
