% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{ldlc}
\alias{ldlc}
\alias{hdlc}
\alias{hdlcse}
\alias{ldlcse}
\alias{trig}
\alias{trigse}
\alias{chdlodds}
\alias{chdloddsse}
\alias{lipid_effect}
\alias{lipid_other}
\alias{lipid_eaf}
\title{Data on lipid effects on coronary artery disease (uncorrelated variants)}
\format{
An object of class \code{numeric} of length 28.

An object of class \code{numeric} of length 28.

An object of class \code{numeric} of length 28.

An object of class \code{numeric} of length 28.

An object of class \code{numeric} of length 28.

An object of class \code{numeric} of length 28.

An object of class \code{numeric} of length 28.

An object of class \code{numeric} of length 28.

An object of class \code{character} of length 28.

An object of class \code{character} of length 28.

An object of class \code{numeric} of length 28.
}
\usage{
ldlc

hdlc

hdlcse

ldlcse

trig

trigse

chdlodds

chdloddsse

lipid_effect

lipid_other

lipid_eaf
}
\description{
Two sets of example data are included in the package: one illustrating uncorrelated variants, and the other correlated variants. These are the data on uncorrelated variants.

The variables \code{ldlc}, \code{hdlc}, \code{trig}, and \code{chdlodds} are the genetic associations with (respectively) LDL-cholesterol, HDL-cholesterol, triglycerides, and coronary heart disease (CHD) risk for 28 genetic variants reported by Waterworth et al (2010). The respective standard errors of the associations are given as \code{ldlcse}, \code{hdlcse}, \code{trigse}, and \code{chdloddsse}.

These data can be used to test out the various functions in the package.
}
\references{
Dawn Waterworth, Sally Ricketts, ..., Manj Sandhu: Genetic variants influencing circulating lipid levels and risk of coronary artery disease. Arterioscler Thromb Vasc Biol 2010; 30:2264-227. doi: 10.1161/atvbaha.109.201020.
}
\keyword{datasets}
