% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_maxlik-methods.R
\name{loglikelihood}
\alias{loglikelihood}
\title{Calculates log-likelihood with uncorrelated variants in two-sample setting (no correlation from sample overlap)}
\usage{
loglikelihood(param, x, sigmax, y, sigmay)
}
\arguments{
\item{param}{Parameters in the model.}

\item{x}{Genetic associations with the exposure.}

\item{sigmax}{Standard errors of associations with the exposure.}

\item{y}{Genetic associations with the outcome.}

\item{sigmay}{Standard errors of genetic associations with the outcome.}
}
\value{
Log-likelihood.
}
\description{
Internal function for calculating log-likelihood for for maximum-likelihood method.
}
\details{
None.
}
\keyword{internal}
