% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_maxlik-methods.R
\name{loglikelihoodcorrel}
\alias{loglikelihoodcorrel}
\title{Calculates log-likelihood with correlated variants in two-sample setting (no correlation from sample overlap)}
\usage{
loglikelihoodcorrel(param, x, Taux, y, Tauy)
}
\arguments{
\item{param}{Parameters in the model.}

\item{x}{Genetic associations with the exposure.}

\item{Taux}{Precision matrix for associations with the exposure.}

\item{y}{Genetic associations with the outcome.}

\item{Tauy}{Precision matrix for associations with the outcome.}
}
\value{
Log-likelihood.
}
\description{
Internal function for calculating log-likelihood for for maximum-likelihood method.
}
\details{
None.
}
\keyword{internal}
