% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_maxlik-methods.R
\name{loglikelihoodrhocorrel}
\alias{loglikelihoodrhocorrel}
\title{Calculates log-likelihood with correlation from sample overlap}
\usage{
loglikelihoodrhocorrel(param, x, y, Tauxy)
}
\arguments{
\item{param}{Parameters in the model.}

\item{x}{Genetic associations with the exposure.}

\item{y}{Genetic associations with the outcome.}

\item{Tauxy}{Precision matrix for associations with the exposure and outcome.}
}
\value{
Log-likelihood.
}
\description{
Internal function for calculating log-likelihood for for maximum-likelihood method.
}
\details{
None.
}
\keyword{internal}

