% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_hetpen-methods.R
\name{model.prior}
\alias{model.prior}
\title{Prior weight function}
\usage{
model.prior(model.size, N.obs, prob.valid.inst)
}
\arguments{
\item{model.size}{Size of subset.}

\item{N.obs}{Number of genetic variants.}

\item{prob.valid.inst}{Probability (prior) of a genetic variant being a valid instrument.}
}
\value{
Prior weight.
}
\description{
Internal function for prior weights.
}
\details{
None.
}
\examples{
model.prior(8, 15, 0.5)

}
\keyword{internal}
