% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.adjust.p.R
\name{mr.adjust.p}
\alias{mr.adjust.p}
\title{Do p value correction on data set}
\usage{
mr.adjust.p(data, column.names, method)
}
\arguments{
\item{data}{Data frame containing column to do p value correction on.}

\item{method}{The method to use for p value correction.}

\item{column.name}{The name of the column to do p value correction on, This must be a list, for it can adjust multiple columns at once}
}
\value{
Adjusted data frame with adjusted pvalues
}
\description{
Do p value correction on data set
}
\examples{
mr.adjust.p()

}
\author{
Matthijs Knigge
}
\keyword{adjust}
\keyword{p}
