% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.allelic.scoring.R
\name{mr.allelic.scoring}
\alias{mr.allelic.scoring}
\title{Allelic Scoring}
\usage{
mr.allelic.scoring(SNP, B, alleles, bfile)
}
\arguments{
\item{SNP}{Vector of SNPs}

\item{B}{Vector of genetic effects}

\item{alleles}{Vector of effect alleles associated with genetic effects}

\item{blife}{absolute path to the bfile}
}
\value{
List with the following elements:
        se: standard error of genetic effect
}
\description{
Allelic Scoring
}
\examples{
mr.allelic.scoring()

}
\author{
Matthijs Knigge
}
\keyword{se}
