% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.calculate.B.R
\name{mr.calculate.B}
\alias{mr.calculate.B}
\title{Calculate effect size}
\usage{
mr.calculate.B(se, Z)
}
\arguments{
\item{se}{Vector of SE on genetic effect}

\item{Z}{Vector of Z scores from genetic effects}
}
\value{
List with the following elements:
        B: genetic effect size
}
\description{
Calculate effect size
}
\examples{
mr.calculate.B()

}
\author{
Matthijs Knigge
}
\keyword{se}
