% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.calculate.se.2.R
\name{mr.calculate.se.2}
\alias{mr.calculate.se.2}
\title{Calculate standard error of effect size}
\usage{
mr.calculate.se.2(p, n, Z)
}
\arguments{
\item{p}{effect allele frequence}

\item{n}{Sample size}

\item{Z}{Vector of Z scores from genetic effects}
}
\value{
List with the following elements:
        se: standard error of genetic effect
}
\description{
Calculate standard error of effect size
}
\examples{
mr.calculate.se.2()

}
\author{
Matthijs Knigge
}
\keyword{se}
