% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.calculate.se.R
\name{mr.calculate.se}
\alias{mr.calculate.se}
\title{Calculate standard error of effect size}
\usage{
mr.calculate.se(B, pval)
}
\arguments{
\item{B}{Vector of genetic effects}

\item{pval}{Vector of p values of genetic effects}
}
\value{
List with the following elements:
        se: standard error of genetic effect
}
\description{
Calculate standard error of effect size
}
\examples{
mr.calculate.se()

}
\author{
Matthijs Knigge
}
\keyword{se}
