% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.chochran.Q.test.R
\name{mr.chochran.Q.test}
\alias{mr.chochran.Q.test}
\title{Perform Chochran's Q test}
\usage{
mr.chochran.Q.test(calculation, method, p.threshold)
}
\arguments{
\item{p.threshold}{p value threshold}
}
\value{
calculation phenotype, SNP, beta.exposure, se.exposure, beta.outcome, se.outcome, beta.iv, beta.iv.se, beta.iv.p

method beta.ivw, beta.ivw.se, beta.ivw.p, beta.egger, beta.egger.se, beta.egger.p, beta.egger.i, beta.egger.i.se, beta.egger.i.p

calculation with removed pleiotrpoic effect, and re-calculated egger, and ivw method
}
\description{
Perform Chochran's Q test
}
\examples{
mr.chochran.Q.test()

}
\author{
Matthijs Knigge
}
\keyword{Q}
\keyword{chochrans}
