% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.clump.R
\name{mr.clump}
\alias{mr.clump}
\title{Perform clumping}
\usage{
mr.clump(data, refdat, clump_kb = 1000, clump_r2 = 0.1, clump_p1 = 5 *
  10^-8, clump_p2 = 5 * 10^-8, verbose = FALSE)
}
\arguments{
\item{data}{data.frame to be clumped. Must contain columns: pval, SNP}

\item{refdat}{path to reference data for clumping (EUR_1000G_phase3_vcf)}

\item{clump_kb}{window size, default = 1000}

\item{clump_r2}{clumping r2 cutoff, default = .1}

\item{clump_p1}{clumping sig level for index SNPs, default = 5*10^-8}

\item{clump_p2}{clumping sig level for index SNPs, default = 5*10^-8}

\item{verbose}{boolean TRUE for clumping and using proxy SNP, FALSE for just removing SNPs in LD. Default: False}
}
\value{
data.frame with removed snps
}
\description{
Perform clumping
}
\examples{
mr.clump(data = data.frame, refdat = '/path/to/refdata/EUR_1000G_phase3_vcf', verbose = TRUE)

}
\author{
Matthijs Knigge
}
\keyword{clump}
