% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.count.n.SNP.R
\name{mr.count.n.SNP}
\alias{mr.count.n.SNP}
\title{Count number of SNPs for methods}
\usage{
mr.count.n.SNP(SNPs.frame, method.frame, column.identifier.phenotype,
  column.identifier.SNP)
}
\arguments{
\item{SNPs.frame}{data frame of SNPs that are used}

\item{method.frame}{summary table of methods used on SNPs}

\item{column.identifier.phenotype}{column name of phenotype identifier}

\item{column.identifier.SNP}{column name of SNP identifier}
}
\value{
Adjusted data frame with number of SNPs added to summary table
}
\description{
Count number of SNPs for methods
}
\examples{
mr.count.n.SNP()

}
\author{
Matthijs Knigge
}
\keyword{adjust}
\keyword{p}
