% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.egger.method.R
\name{mr.egger.method}
\alias{mr.egger.method}
\title{Perform mendelian randomization-egger method}
\usage{
mr.egger.method(By, Bx, By.se, Bx.se)
}
\arguments{
\item{By}{Vector of genetic effects of outcome}

\item{Bx}{Vector of genetic effects of exposure}

\item{By.se}{Standard errors of genetic effects on outcome}

\item{Bx.se}{Standard errors of genetic effects on exposure}
}
\value{
List of with the following elements:
        egger: MR estimate slope
        egger.se: Standard error of MR estimate
        egger.p: p-value of MR estimate
        egger.i: Estimate of horizontal pleiotropy (intercept)
        egger.i.se: Standard error of intercept
        egger.i.p: p-value of intercept
}
\description{
Perform mendelian randomization-egger method
}
\examples{
mr.egger.method()

}
\author{
Matthijs Knigge
}
\keyword{egger}
\keyword{mendelian}
\keyword{method}
\keyword{randomization}
