% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.filter.significance.R
\name{mr.filter.significance}
\alias{mr.filter.significance}
\title{Filter on pvalue}
\usage{
mr.filter.significance(summary.table, column.ivw.fdr, column.egger.fdr,
  threshold)
}
\arguments{
\item{summary.table}{data frame summarizing all methods used}

\item{column.ivw.fdr}{column name of pvalues to be filtered for ivw}

\item{column.egger.fdr}{column name of pvalues to be filtered for egger}

\item{threshold}{threshold for filtering pvalues}
}
\value{
Adjusted data frame with filtered on threshold
}
\description{
Filter on pvalue
}
\examples{
mr.filter.significance()

}
\author{
Matthijs Knigge
}
\keyword{adjust}
\keyword{p}
