% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.find.missing.allelic.information.R
\name{mr.find.missing.allelic.information}
\alias{mr.find.missing.allelic.information}
\title{Find missing allele}
\usage{
mr.find.missing.allelic.information(data, thousand.G)
}
\arguments{
\item{data}{data.frame that must contain effect_allele, and SNP columns}

\item{thousand.G}{absolute location to 1000Genomes file}
}
\value{
data with both alleles effect_allele and other_allele
}
\description{
If the outcome or exposure is missing allelic information, it will sort out what is the missing allele, and what is the effect and other allele.
}
\examples{
mr.find.missing.allelic.information(data = outcome, thousand.G = '/path/to/thousand/genomes/file.bim')

}
\author{
Matthijs Knigge
}
\keyword{allele}
\keyword{missing}
