% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.forest.plot.R
\name{mr.forest.plot}
\alias{mr.forest.plot}
\title{Plot forest}
\usage{
mr.forest.plot(SNP, iv, iv.se, chochran.Q = NULL, ivw = NULL,
  ivw.se = NULL, egger = NULL, egger.se = NULL, ivw.Q = NULL,
  ivw.Q.se = NULL, egger.Q = NULL, egger.Q.se = NULL, outcome.name,
  exposure.name)
}
\arguments{
\item{iv}{vector of causal effect estimate}

\item{iv.se}{vector of standard error of causal effect estimate}

\item{chochran.Q}{vector indicating if the SNPs is removed by Chochran's Q test due to pleiotropic effects. Default is NULL.}

\item{ivw}{the beta estimate of inverse variance weighted method}

\item{ivw.se}{the standard error of the inverse variance weighted method}

\item{egger}{the beta estimate of egger method}

\item{egger.se}{Standard error of MR egger estimate}

\item{ivw.Q}{numeric after correcting for pleotropic effect this is the re-calculated inverse variance weighted estimate. Default is NULL.}

\item{ivw.Q.se}{numeric after correcting for pleotropic effect this is the re-calculated standard error of the inverse variance weighted estimate. Default is NULL.}

\item{egger.Q}{numeric after correcting for pleotropic effect this is the re-calculated egger estimate. Default is NULL.}

\item{egger.Q.se}{numeric after correcting for pleotropic effect this is the re-calculated standard error of the egger estimate. Default is NULL.}
}
\value{
forest plot
}
\description{
Plot forest
}
\examples{
mr.forest.plot()

}
\author{
Matthijs Knigge
}
\keyword{forest}
