% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.funnel.plot.R
\name{mr.funnel.plot}
\alias{mr.funnel.plot}
\title{MR funnel plot}
\usage{
mr.funnel.plot(iv, iv.se)
}
\arguments{
\item{iv}{vector of causal effect estimate}

\item{iv.se}{vector of standard error of causal effect estimate}
}
\value{
funnel plot
}
\description{
MR funnel plot
}
\examples{
mr.funnel.plot()

}
\author{
Matthijs Knigge
}
\keyword{funnel}
