% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.get.chr.pos.R
\name{mr.get.chr.pos}
\alias{mr.get.chr.pos}
\title{Get chromosome and position based on reference file}
\usage{
mr.get.chr.pos(data, refdata)
}
\arguments{
\item{data}{data.frame must contain column SNP}

\item{refdata}{absolute path to reference data}
}
\value{
data with chr and pos added
}
\description{
Get chromosome and position based on reference file
}
\examples{
mr.get.chr.pos()

}
\author{
Matthijs Knigge
}
\keyword{chromosome}
\keyword{get}
\keyword{position}
