% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.inverse.variance.weighted.method.R
\name{mr.inverse.variance.weighted.method}
\alias{mr.inverse.variance.weighted.method}
\title{Perform inverse variance weighted method}
\usage{
mr.inverse.variance.weighted.method(By, Bx, By.se, Bx.se, subset = NULL)
}
\arguments{
\item{By}{Vector of genetic effects of outcome}

\item{Bx}{Vector of genetic effects of exposure}

\item{By.se}{Standard errors of genetic effects on outcome}

\item{Bx.se}{Standard errors of genetic effects on exposure}

\item{subset}{vector containing 1-0 indicating to use it in the calculation or not. Default is NULL.}
}
\value{
List with the following elements:
        ivw: MR estimate
        ivw.se: standard error
        ivw.p: p-value
}
\description{
Perform inverse variance weighted method
}
\examples{
mr.inverse.variance.weighted.method()

}
\author{
Matthijs Knigge
}
\keyword{inverse}
\keyword{method}
\keyword{variance}
\keyword{weighted}
