% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.clump.R
\name{mr.parse.verbose}
\alias{mr.parse.verbose}
\title{Parse verbose, get proxy snps}
\usage{
mr.parse.verbose(file, SNPs.list.1, SNPs.list.2)
}
\arguments{
\item{file}{file conatianing snps and pval to clump}

\item{SNPs}{list of snps to compare with proxy determination}
}
\value{
list of SNPs
}
\description{
Parse verbose, get proxy snps
}
\author{
Matthijs Knigge
}
