% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.plot.R
\name{mr.plot}
\alias{mr.plot}
\title{Plot MR results}
\usage{
mr.plot(By, Bx, By.se, Bx.se, iv, iv.se)
}
\arguments{
\item{By}{Vector of genetic effects on exposure}

\item{Bx}{Vector of genetic effects on outcome}

\item{By.se}{Standard errors of genetic effects on exposure}

\item{Bx.se}{Standard errors of genetic effects on outcome}

\item{iv}{causal effect estimate}

\item{iv.se}{standard error of causal effect estimate}
}
\value{
plot object
}
\description{
Plot MR results
}
\author{
Matthijs Knigge
}
