% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.pre.process.R
\name{mr.pre.process}
\alias{mr.pre.process}
\title{Perform pre-process}
\usage{
mr.pre.process(B, B.se, pval, effect_allele, other_allele, SNP)
}
\arguments{
\item{B}{Vector of genetic effects}

\item{B.se}{Standard errors of genetic effects}

\item{pval}{p-value}

\item{effect_allele}{vector containing the effect_alleles}

\item{other_allele}{vector containing the effect_alleles}

\item{SNP}{vector containing rs_ids}
}
\value{
cleaned data
}
\description{
pre-process filters on genome-wide significance, SNPs without effectsize, no allelic information, removal of duplicates, removes alleles from which it is not possible to measure the direction, flips alleles and beta when needed,
}
\examples{
mr.pre.process()

}
\author{
Matthijs Knigge
}
\keyword{pre}
\keyword{process}
