% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.qq.plot.R
\name{mr.qq.plot}
\alias{mr.qq.plot}
\title{Make quantile-quantile plot for inverse variance weighted method and MR-egger method}
\usage{
mr.qq.plot(By, Bx, By.se, Bx.se, egger = FALSE, ivw = FALSE,
  chochran.Q = NULL, outcome.name, exposure.name)
}
\arguments{
\item{By}{Vector of genetic effects from outcome}

\item{Bx}{Vector of genetic effects from exposure}

\item{By.se}{Vector of standard errors from genetic effect of outcome}

\item{Bx.se}{Vector of standard errors from genetic effect of exposure}

\item{egger}{boolean indicating if the residuals from this method should be plotted. Default FALSE}

\item{ivw}{boolean indicating if the residuals from this method should be plotted. Default FALSE}

\item{chochran.Q}{vector containing 0-1 indicating which observations where not used in recalculating the effect. Default NULL.}

\item{outcome.name}{name of the outcome}

\item{exposure.name}{name of the exposure}
}
\value{
qqplot
}
\description{
Make quantile-quantile plot for inverse variance weighted method and MR-egger method
}
\examples{
mr.qq.plot()

}
\author{
Matthijs Knigge
}
\keyword{qqplot}
