% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.remove.region.R
\name{mr.remove.region}
\alias{mr.remove.region}
\title{Remove region from file}
\usage{
mr.remove.region(chr, lower.limit, upper.limit, data)
}
\arguments{
\item{chr}{chrosome number from wich the region must be romved}

\item{lower.limit}{lower limit of the region to be removed in kb}

\item{upper.limit}{upper limit of the region to be removed in kb}

\item{data}{data.frame must contain column pos, SNP, and chr}
}
\value{
data without region selected
}
\description{
Remove region from file
}
\examples{
mr.remove.region()

}
\author{
Matthijs Knigge
}
\keyword{region}
\keyword{remove}
