% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.show.tag.R
\name{mr.show.tag}
\alias{mr.show.tag}
\title{clumping show-tag for finding proxy SNPs}
\usage{
mr.show.tag(SNP, refdat, tag_r2 = 0.1, tag_kb = 1000)
}
\arguments{
\item{SNP}{vector containing SNPs to be proxied}

\item{refdat}{path to reference data for tagging (EUR_1000G_phase3_vcf)}

\item{tag_r2}{clumping r2 cutoff, default = .1}

\item{tag_kb}{window size, default = 1000}
}
\value{
data.frame with removed snps
}
\description{
clumping show-tag for finding proxy SNPs
}
\examples{
mr.show.tag(SNP = SNP, refdat = '/path/to/refdata/EUR_1000G_phase3_vcf')

}
\author{
Matthijs Knigge
}
\keyword{show}
\keyword{tag}
