% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MRAll-class}
\alias{MRAll-class}
\title{MRAll Class}
\description{
An object containing the estimates produced using the \code{mr_allmethods} function.
}
\section{Slots}{

\describe{
\item{\code{Data}}{The \code{mr_input} object that was used as an input to the \code{mr_allmethods} function. This includes the original data, so that a call to \code{mr_plot} can plot the original data and the various causal estimates.}

\item{\code{Values}}{A data.frame object comprising estimates from the various methods called by the \code{mr_allmethods} function. The first column gives the names of the methods, then the causal estimates, standard errors, 95\% confidence intervals, and p-values.}

\item{\code{Method}}{A string indicating whether all methods are implemented (\code{"all"}, the default option), or just main methods (\code{"main"}), or only a subset of methods (\code{"ivw"}, \code{"egger"}, or \code{"median"}).}
}}

