% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MRcML-class}
\alias{MRcML-class}
\title{MRcML Class}
\description{
An object containing the results of MRcML.
}
\section{Slots}{

\describe{
\item{\code{Exposure}}{The names of the exposure variables.}

\item{\code{Outcome}}{The name of the outcome variable.}

\item{\code{Estimate}}{Estimate of theta.}

\item{\code{StdError}}{Standard error of estimate.}

\item{\code{Pvalue}}{p-value of estimate.}

\item{\code{BIC_invalid}}{Set of selected invalid IVs if cML-BIC is performed, i.e. without MA or DP.}

\item{\code{GOF1_p}}{p-value of the first goodness-of-fit test.}

\item{\code{GOF2_p}}{p-value of the second goodness-of-fit test.}

\item{\code{SNPs}}{The number of SNPs that were used in the calculation.}

\item{\code{Alpha}}{Significance level for the confidence interval for estimate, default is 0.05.}

\item{\code{CILower}}{Lower bound of the confidence interval for estimate.}

\item{\code{CIUpper}}{Upper bound of the confidence interval for estimate.}

\item{\code{MA}}{Indicator of whether model average is applied.}

\item{\code{DP}}{Indicator of whether data perturbation is applied.}
}}

