% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MRHetPen-class}
\alias{MRHetPen-class}
\title{MRHetPen Class}
\description{
An object containing the estimate produced using the heterogeneity-penalized model-averaging mode-based estimation method as well as various statistics.
}
\section{Slots}{

\describe{
\item{\code{Exposure}}{The names of the exposure variables.}

\item{\code{Outcome}}{The name of the outcome variable.}

\item{\code{Prior}}{The value of the prior probability of a genetic variant being a valid instrument (default is 0.5).}

\item{\code{Estimate}}{The causal estimate from the heterogeneity-penalized method.}

\item{\code{CIRange}}{The confidence interval for \code{Estimate} based on a grid search.}

\item{\code{CILower}}{The lower limit of the confidence interval. If the confidence interval contains multiple ranges, then lower limits of all ranges will be reported.}

\item{\code{CIUpper}}{The upper limit of the confidence interval. If the confidence interval contains multiple ranges, then upper limits of all ranges will be reported.}

\item{\code{CIMin}}{The smallest value used in the search to find the confidence interval.}

\item{\code{CIMax}}{The largest value used in the search to find the confidence interval.}

\item{\code{CIStep}}{The step size used in the search to find the confidence interval.}

\item{\code{Alpha}}{The significance level used in constructing the confidence interval (default is 0.05).}

\item{\code{SNPs}}{The number of SNPs that were used in the calculation.}
}}

