% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_mvivwme-methods.R
\name{mv_norm}
\alias{mv_norm}
\title{Sampling from multivariate normal distribution}
\usage{
mv_norm(n, mu, Sigma)
}
\arguments{
\item{n}{Number of draws.}

\item{mu}{Mean vector.}

\item{Sigma}{Covariance matrix.}
}
\value{
Random vector of length n.
}
\description{
Sampling from multivariate normal distribution
}
\details{
None.
}
\examples{
mv_norm(1, 0.2, matrix(0.1))

}
\keyword{internal}
