% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_mvcML-methods.R
\name{MVcML_SdTheta}
\alias{MVcML_SdTheta}
\title{Standard error estimate for MVMR-cML-BIC}
\usage{
MVcML_SdTheta(b_exp, b_out, Sig_inv_l, theta, zero_ind, r_vec = NULL)
}
\arguments{
\item{b_exp}{A matrix of SNP effects on the exposure variable.}

\item{b_out}{A vector of SNP effects on the outcome variable.}

\item{Sig_inv_l}{A list of inverse of covariance matrix.}

\item{theta}{A vector of final estimates of causal effect of each exposure by MVMR-cML-BIC obtained from \code{MVmr_cML_DP}.}

\item{zero_ind}{A vector of the index of valid IVs.}

\item{r_vec}{A vector of estimated horizontal pleiotropic effects.}
}
\value{
A vector
}
\description{
This is based on the profile likelihood of the set of valid IVs, which is not robust to uncertainty in model selection.
}
