% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MVMedian-class}
\alias{MVMedian-class}
\title{MRMVMedian class}
\description{
An object containing the estimates produced using the multivariable median method as well as various statistics.
}
\section{Slots}{

\describe{
\item{\code{Exposure}}{The names of the exposure variables.}

\item{\code{Outcome}}{The name of the outcome variable.}

\item{\code{Estimate}}{The causal estimates from the multivariable median method.}

\item{\code{StdError}}{The standard errors associated with \code{Estimate} (obtained from bootstrapping).}

\item{\code{CILower}}{The lower bounds of the confidence intervals for \code{Estimate} based on \code{StdError}.}

\item{\code{CIUpper}}{The upper bounds of the confidence intervals for \code{Estimate} based on \code{StdError}.}

\item{\code{Alpha}}{The significance level used in constructing the confidence interval (default is 0.05).}

\item{\code{Pvalue}}{P-values associated with the causal estimates from the Wald method.}

\item{\code{SNPs}}{The number of SNPs that used in the calculation.}
}}

