% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MVMRcML-class}
\alias{MVMRcML-class}
\title{MVMRcML Class}
\description{
An object containing the results of MVMRcML.
}
\section{Slots}{

\describe{
\item{\code{Exposure}}{The names of the exposure variables.}

\item{\code{Outcome}}{The name of the outcome variable.}

\item{\code{Estimate}}{The causal estimates from the multivariable MRcML method.}

\item{\code{StdError}}{The standard errors associated with \code{Estimate}.}

\item{\code{CILower}}{The lower bounds of the confidence intervals for \code{Estimate} based on \code{StdError}.}

\item{\code{CIUpper}}{The upper bounds of the confidence intervals for \code{Estimate} based on \code{StdError}.}

\item{\code{Alpha}}{The significance level used in constructing the confidence interval (default is 0.05).}

\item{\code{Pvalue}}{P-values associated with the causal estimates from the multivariable MRcML method.}

\item{\code{BIC_invalid}}{The index set of selected invalid IVs by MVMRcML-BIC.}

\item{\code{K_hat}}{The number of selected invalid IVs by MVMRcML-BIC, or a vector for each data perturbation in MVMRcML-DP.}

\item{\code{eff_DP_B}}{The number of data perturbations with successful convergence in MVMRcML-DP.}

\item{\code{SNPs}}{The number of SNPs that were used in the calculation.}

\item{\code{DP}}{Indicator of whether data perturbation is applied.}
}}

