% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_ivw-methods.R
\name{penalised.weights}
\alias{penalised.weights}
\title{Calculates p-values for penalization of weights}
\usage{
penalised.weights(.Bx, .Bxse, .By, .Byse)
}
\arguments{
\item{.Bx}{Beta-coefficient for genetic association with the risk factor.}

\item{.Bxse}{Standard error of genetic association with the risk factor.}

\item{.By}{Beta-coefficient for genetic association with the outcome.}

\item{.Byse}{Standard error of genetic association with the outcome.}
}
\value{
P-values corresponding to heterogeneity test for each genetic variant in turn (based on a chi-squared-1 distribution).
}
\description{
Internal function for calculating penalized weights in conjunction with \code{r.weights}.

These weights are used in either the \code{mr_ivw} or \code{mr_egger} functions when \code{penalized = TRUE}, or in the \code{mr_median} function when \code{method = "penalized"}.
}
\details{
None.
}
\examples{
penalised.weights(ldlc, ldlcse, chdlodds, chdloddsse)

}
\keyword{internal}
