% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_mvcML-methods.R
\name{pl}
\alias{pl}
\title{Profile likelihood of valid IVs}
\usage{
pl(x, b_exp_v, b_out_v, Sig_inv_v)
}
\description{
calculate the profile likelihood of valid IVs up to a constant
}
\keyword{internal}
