% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtraFunctions.R
\name{simpleCap}
\alias{simpleCap}
\title{Capitalize a word}
\usage{
simpleCap(x)
}
\arguments{
\item{x}{A word.}
}
\value{
Returns a capitalized word.
}
\description{
Internal function for capitalizing a word.
}
\details{
None.
}
\examples{
simpleCap(x = "weighted")

}
\keyword{internal}

