% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_median-methods.R
\name{weighted.median}
\alias{weighted.median}
\title{Weighted median function}
\usage{
weighted.median(theta, weights)
}
\arguments{
\item{theta}{Ratio causal estimates for each genetic variant considered individually.}

\item{weights}{Weights.}
}
\value{
Causal estimate.
}
\description{
Internal function for calculating weighted median estimate (or simple median estimate if weights are all equal).
}
\details{
None.
}
\examples{
weighted.median(theta = chdlodds/ldlc, weights = chdloddsse/ldlc)

}
\keyword{internal}

