% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter-methods.R
\docType{methods}
\name{getter}
\alias{getter}
\alias{$,WeightedMedian-method}
\alias{$,IVW-method}
\alias{$,MRMBE-method}
\alias{$,Egger-method}
\alias{$,MRAll-method}
\alias{$,MRInput-method}
\alias{$,MaxLik-method}
\alias{$,MVIVW-method}
\alias{$,MVEgger-method}
\alias{$,MRMVInput-method}
\alias{$,MRHetPen-method}
\alias{$,MRConMix-method}
\alias{$,MVMedian-method}
\alias{$,MVLasso-method}
\alias{$,MRLasso-method}
\alias{$,DIVW-method}
\alias{$,MRcML-method}
\alias{$,PIVW-method}
\alias{$,MVMRcML-method}
\title{Applies method $ to different classes}
\usage{
\S4method{$}{WeightedMedian}(x, name)

\S4method{$}{IVW}(x, name)

\S4method{$}{MRMBE}(x, name)

\S4method{$}{Egger}(x, name)

\S4method{$}{MRAll}(x, name)

\S4method{$}{MRInput}(x, name)

\S4method{$}{MaxLik}(x, name)

\S4method{$}{MVIVW}(x, name)

\S4method{$}{MVEgger}(x, name)

\S4method{$}{MRMVInput}(x, name)

\S4method{$}{MRHetPen}(x, name)

\S4method{$}{MRConMix}(x, name)

\S4method{$}{MVMedian}(x, name)

\S4method{$}{MVLasso}(x, name)

\S4method{$}{MRLasso}(x, name)

\S4method{$}{DIVW}(x, name)

\S4method{$}{MRcML}(x, name)

\S4method{$}{PIVW}(x, name)

\S4method{$}{MVMRcML}(x, name)
}
\arguments{
\item{x}{Object.}

\item{name}{Name of slot.}
}
\description{
Enables slots of objects in this package to be referenced easily.
}
\keyword{internal}
