% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.chochran.Q.test.R
\name{mr.chochran.Q.test}
\alias{mr.chochran.Q.test}
\title{Perform Chochran's Q test}
\usage{
mr.chochran.Q.test(data, pval)
}
\arguments{
\item{data}{data.frame containen SNP, By, Bx, By.se, Bx.se, iv, iv.se}

\item{pval}{p value threshold}
}
\value{
data with column chochran's Q column added which says if the SNPs is used or not in the method.
}
\description{
Chochran's Q is a iterative method that tries to find pleiotropic SNPs, and filters those SNPs out.
}
\examples{
mr.chochran.Q.test()

}
\author{
Matthijs Knigge
}
\keyword{Q}
\keyword{chochrans}
