% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.Pascal.R
\name{mr.Pascal}
\alias{mr.Pascal}
\title{Pascal (Pathway scoring algorithm)}
\usage{
mr.Pascal(SNPs = NULL, pval = NULL)
}
\arguments{
\item{SNPs}{vector of SNPs. Default is NULL.}

\item{pval}{vector p values associated with the SNPs. Default is NULL.}
}
\value{
List with the following elements:
}
\description{
Pascal is a tool that allows gene and pathway-level analysis of GWAS association results without the need to access genotypic data. This is an executable that can not be exported. This must be downloaded and installed by the user. Pascal is available for download here: https://www2.unil.ch/cbg/index.php?title=Pascal. Individual SNPs are mapped to genes and their association p-values are combined into gene scores; second, genes are grouped into pathways and their gene scores are combined into pathway scores.
}
\examples{
mr.Pascal()

}
\author{
Matthijs Knigge
}
\keyword{Pascal}
