% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.plot.R
\name{mr.plot}
\alias{mr.plot}
\title{Plot MR results}
\usage{
mr.plot(By, Bx, By.se, Bx.se, iv, iv.se, ivw = NULL, egger = NULL,
  egger.i = NULL, chochran.Q = NULL, ivw.Q = NULL, egger.Q = NULL,
  egger.i.Q = NULL, outcome.name, exposure.name, legend = TRUE)
}
\arguments{
\item{By}{Vector of genetic effects of outcome}

\item{Bx}{Vector of genetic effects of exposure}

\item{By.se}{Standard errors of genetic effects on outcome}

\item{Bx.se}{Standard errors of genetic effects on exposure}

\item{iv}{vector of causal effect estimate}

\item{iv.se}{vector of standard error of causal effect estimate}

\item{ivw}{numeric causal estimate inverse variance weighted method. Default is NULL.}

\item{egger}{numeric causal estimate of egger method, this is the slope. Default is NULL.}

\item{egger.i}{numeric pleiotropic estimate of egger, this is the intersept. Default is NULL.}

\item{chochran.Q}{vector indicating if the SNPs is removed by Chochran's Q test due to pleiotropic effects. Default is NULL.}

\item{ivw.Q}{numeric after correcting for pleotropic effect this is the re-calculated inverse variance weighted estimate. Default is NULL.}

\item{egger.Q}{numeric after correcting for pleotropic effect this is the re-calculated egger estimate. Default is NULL.}

\item{egger.i.Q}{numeric after correcting for pleotropic effect this is the re-calculated egger pleiotropic estimate. Default is NULL.}

\item{outcome.name}{string of the outcome name. Default is NULL.}

\item{exposure.name}{string of the exposure name. Default is NULL.}

\item{legend}{boolean use show legend, or do not. Default is TRUE}
}
\value{
plot object
}
\description{
Plot MR results
}
\examples{
mr.plot <- function(By, Bx, By.se, Bx.se, iv, iv.se, ivw = NULL, egger = NULL, egger.i = NULL, chochran.Q = NULL,
                    ivw.Q = NULL, egger.Q = NULL, egger.i.Q = NULL, outcome.name, exposure.name, legend = TRUE)
}
\author{
Matthijs Knigge
}
\keyword{mr.plot}
