% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.mr.qq.p.distribution.R
\name{mr.qq.p.distribution}
\alias{mr.qq.p.distribution}
\title{MR funnel plot}
\usage{
mr.qq.p.distribution(egger = NULL, ivw = NULL, egger.Q = NULL,
  ivw.Q = NULL, CI = 0.95)
}
\arguments{
\item{egger}{vector of pvalues from egger estimate. Default NULL.}

\item{ivw}{vector of pvalues from ivw estimate. Default NULL.}

\item{egger.Q}{vector of pvalues from egger estimate after Chochren's Q test. Default NULL.}

\item{ivw.Q}{vector of pvalues from ivw estimate after Chochren's Q test. Default NULL.}

\item{CI}{confidence interval}
}
\value{
funnel plot
}
\description{
MR funnel plot
}
\examples{
mr.qq.p.distribution()

}
\author{
Matthijs Knigge
}
\keyword{qqplot}
