% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values-methods.R
\docType{methods}
\name{values}
\alias{values}
\alias{values,WeightedMedian-method}
\alias{values,IVW-method}
\alias{values,Egger-method}
\title{Applies method values() to different output classes}
\usage{
\S4method{values}{WeightedMedian}(object)

\S4method{values}{IVW}(object)

\S4method{values}{Egger}(object)
}
\arguments{
\item{object}{Object (could be an object of class "WeightedMedian", "Egger", or "IVW").}
}
\description{
Enables the internal function \code{values}, used in the \code{mr_allmethods} function.
}
\keyword{internal}
